refactor: move core, js and preset-algolia types to shared
