+++ packages/autocomplete-core/src/types/index.ts
@@ -1,13 +1,3 @@
-export * from './AutocompleteApi';
-export * from './AutocompleteCollection';
-export * from './AutocompleteContext';
-export * from './AutocompleteEnvironment';
-export * from './AutocompleteOptions';
-export * from './AutocompleteSource';
-export * from './AutocompletePropGetters';
-export * from './AutocompletePlugin';
-export * from './AutocompleteReshape';
-export * from './AutocompleteSetters';
-export * from './AutocompleteState';
+export * from '@algolia/autocomplete-shared/dist/esm/core';
 export * from './AutocompleteStore';
 export * from './AutocompleteSubscribers';
