+++ packages/autocomplete-js/src/types/index.ts
@@ -1,14 +1,3 @@
+export * from '@algolia/autocomplete-shared/dist/esm/js';
 export * from './AutocompleteApi';
-export * from './AutocompleteClassNames';
-export * from './AutocompleteCollection';
-export * from './AutocompleteComponents';
 export * from './AutocompleteDom';
-export * from './AutocompleteOptions';
-export * from './AutocompletePlugin';
-export * from './AutocompletePropGetters';
-export * from './AutocompleteRender';
-export * from './AutocompleteRenderer';
-export * from './AutocompleteSource';
-export * from './AutocompleteState';
-export * from './AutocompleteTranslations';
-export * from './HighlightHitParams';
