+++ packages/autocomplete-plugin-algolia-insights/src/types/EventParams.ts
@@ -1,4 +1,4 @@
-import { AutocompleteState } from '@algolia/autocomplete-js';
+import { AutocompleteState } from '@algolia/autocomplete-shared';
 
 import {
   ClickedObjectIDsAfterSearchParams,
