+++ packages/autocomplete-preset-algolia/package.json
@@ -34,11 +34,9 @@
     "@algolia/autocomplete-shared": "1.8.3"
   },
   "devDependencies": {
-    "@algolia/client-search": "4.16.0",
     "algoliasearch": "4.16.0"
   },
   "peerDependencies": {
-    "@algolia/client-search": ">= 4.9.1 < 6",
     "algoliasearch": ">= 4.9.1 < 6"
   }
 }
