+++ packages/autocomplete-preset-algolia/src/requester/getAlgoliaFacets.ts
@@ -1,7 +1,7 @@
 import type { MultipleQueriesQuery } from '../types';
+import { RequestParams } from '../types';
 
 import { createAlgoliaRequester } from './createAlgoliaRequester';
-import { RequestParams } from './createRequester';
 
 /**
  * Retrieves Algolia facet hits from multiple indices.
