+++ packages/autocomplete-preset-algolia/src/types/index.ts
@@ -1,1 +1,2 @@
-export * from './algoliasearch';
+export * from '@algolia/autocomplete-shared/dist/esm/preset-algolia/algoliasearch';
+export * from '@algolia/autocomplete-shared/dist/esm/preset-algolia/createRequester';
