+++ packages/autocomplete-shared/package.json
@@ -25,5 +25,13 @@
     "on:change": "concurrently \"yarn build:esm\" \"yarn build:types\"",
     "prepare": "yarn build:esm && yarn build:types",
     "watch": "watch \"yarn on:change\" --ignoreDirectoryPattern \"/dist/\""
+  },
+  "devDependencies": {
+    "@algolia/client-search": "4.16.0",
+    "algoliasearch": "4.16.0"
+  },
+  "peerDependencies": {
+    "@algolia/client-search": ">= 4.9.1 < 6",
+    "algoliasearch": ">= 4.9.1 < 6"
   }
 }
