+++ packages/autocomplete-shared/src/core/AutocompleteEnvironment.ts
