+++ packages/autocomplete-shared/src/core/AutocompleteNavigator.ts
