+++ packages/autocomplete-shared/src/core/AutocompleteOptions.ts
@@ -1,5 +1,4 @@
-import { CreateAlgoliaInsightsPluginParams } from '@algolia/autocomplete-plugin-algolia-insights';
-import { MaybePromise } from '@algolia/autocomplete-shared';
+import type { MaybePromise } from '../MaybePromise';
 
 import { AutocompleteScopeApi, BaseItem } from './AutocompleteApi';
 import { AutocompleteEnvironment } from './AutocompleteEnvironment';
@@ -177,15 +176,8 @@ export interface AutocompleteOptions<TItem extends BaseItem> {
    * @link https://www.algolia.com/doc/ui-libraries/autocomplete/api-reference/autocomplete-js/autocomplete/#param-reshape
    */
   reshape?: Reshape<TItem>;
-  /**
-   * Whether to enable the Insights plugin and load the Insights library if it has not been loaded yet.
-   *
-   * See [**autocomplete-plugin-algolia-insights**](https://www.algolia.com/doc/ui-libraries/autocomplete/api-reference/autocomplete-plugin-algolia-insights/) for more information.
-   *
-   * @default false
-   * @link https://www.algolia.com/doc/ui-libraries/autocomplete/api-reference/autocomplete-js/autocomplete/#param-insights
-   */
-  insights?: CreateAlgoliaInsightsPluginParams | boolean;
+
+  insights?: boolean;
 }
 
 // Props manipulated internally with default values.
