+++ packages/autocomplete-shared/src/core/AutocompleteSource.ts
@@ -1,10 +1,11 @@
+import type { FacetHit, Hit } from '@algolia/client-search';
+
+import type { MaybePromise } from '../MaybePromise';
 import type {
   SearchForFacetValuesResponse,
   SearchResponse,
-  RequesterDescription,
-} from '@algolia/autocomplete-preset-algolia';
-import type { MaybePromise } from '@algolia/autocomplete-shared';
-import { FacetHit, Hit } from '@algolia/client-search';
+} from '../preset-algolia/algoliasearch';
+import type { RequesterDescription } from '../preset-algolia/createRequester';
 
 import { AutocompleteScopeApi, BaseItem } from './AutocompleteApi';
 import { GetSourcesParams } from './AutocompleteOptions';
