+++ packages/autocomplete-shared/src/core/index.ts
@@ -1,0 +1,11 @@
+export * from './AutocompleteApi';
+export * from './AutocompleteCollection';
+export * from './AutocompleteContext';
+export * from './AutocompleteEnvironment';
+export * from './AutocompleteOptions';
+export * from './AutocompleteSource';
+export * from './AutocompletePropGetters';
+export * from './AutocompletePlugin';
+export * from './AutocompleteReshape';
+export * from './AutocompleteSetters';
+export * from './AutocompleteState';
