+++ packages/autocomplete-shared/src/index.ts
@@ -14,3 +14,4 @@ export * from './UserAgent';
 export * from './userAgents';
 export * from './version';
 export * from './warn';
+export * from './js';
