+++ packages/autocomplete-shared/src/js/AutocompleteTranslations.ts
