+++ packages/autocomplete-shared/src/js/HighlightHitParams.ts
