+++ packages/autocomplete-shared/src/js/index.ts
@@ -1,0 +1,12 @@
+export * from './AutocompleteClassNames';
+export * from './AutocompleteCollection';
+export * from './AutocompleteComponents';
+export * from './AutocompleteOptions';
+export * from './AutocompletePlugin';
+export * from './AutocompletePropGetters';
+export * from './AutocompleteRender';
+export * from './AutocompleteRenderer';
+export * from './AutocompleteSource';
+export * from './AutocompleteState';
+export * from './AutocompleteTranslations';
+export * from './HighlightHitParams';
