+++ packages/autocomplete-shared/src/preset-algolia/algoliasearch.ts
